refactor: simplify authorization logic by replacing try-catch blocks with conditional checks in API controllers

+++ app/Http/Controllers/Api/V1/CustomerOrdersController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\Api\V1\UpdateOrderRequest;
 use App\Http\Resources\V1\OrderResource;
 use App\Models\Order;
 use App\Policies\V1\OrderPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class CustomerOrdersController extends ApiController
@@ -31,16 +30,13 @@ class CustomerOrdersController extends ApiController
      */
     public function store(StoreOrderRequest $request, $customer_id)
     {
-        try {
-            // policy
-             $this->isAble('store', Order::class);
-
-             return new OrderResource(Order::create($request->mappedAttributes([
+        if ($this->isAble('store', Order::class)) {
+            return new OrderResource(Order::create($request->mappedAttributes([
                 'customer' => 'user_id'
-             ])));
-         } catch (AuthorizationException $ex) {
-             return $this->error('You are not authorized to create that resource', 403);
-         }
+            ])));
+        }
+
+        return $this->error('You are not authorized to create that resource', 403);
     }
 
     /**
@@ -54,15 +50,15 @@ class CustomerOrdersController extends ApiController
                             ->where('user_id', $customer_id)
                             ->firstOrFail();
 
-            $this->isAble('replace', $order);
+            if ($this->isAble('replace', $order)) {
+                $order->update($request->mappedAttributes());
+                return new OrderResource($order);
+            }
 
-            $order->update($request->mappedAttributes());
-            return new OrderResource($order);
+            return $this->error('You are not authorized to update that resource', 403);
 
         } catch (ModelNotFoundException $exception) {
             return $this->error('Order cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 403);
         }
     }
 
@@ -77,14 +73,14 @@ class CustomerOrdersController extends ApiController
                             ->where('user_id', $customer_id)
                             ->firstOrFail();
 
-            $this->isAble('update', $order);
+            if ($this->isAble('update', $order)) {
+                $order->update($request->mappedAttributes());
+                return new OrderResource($order);
+            }
 
-            $order->update($request->mappedAttributes());
-            return new OrderResource($order);
+            return $this->error('You are not authorized to update that resource', 403);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Order cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 403);
         }
     }
 
@@ -98,14 +94,14 @@ class CustomerOrdersController extends ApiController
                             ->where('user_id', $customer_id)
                             ->firstOrFail();
 
-            $this->isAble('delete', $order);
+            if ($this->isAble('delete', $order)) {
+                $order->delete();
+                return $this->ok('Order successfully deleted');
+            }
 
-            $order->delete();
-            return $this->ok('Order successfully deleted');
+            return $this->error('You are not authorized to delete that resource', 403);
         } catch (ModelNotFoundException $exception) {
             return $this->error('Order cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to delete that resource', 403);
         }
     }
 }
