+++ app/Http/Controllers/Api/V1/UserController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\Api\V1\StoreUserRequest;
 use App\Http\Requests\Api\V1\UpdateUserRequest;
 use App\Http\Resources\V1\UserResource;
 use App\Policies\V1\UserPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class UserController extends ApiController
@@ -30,14 +29,11 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            // policy
-             $this->isAble('store', User::class);
+        if ($this->isAble('store', User::class)) {
+            return new UserResource(User::create($request->mappedAttributes()));
+        }
 
-             return new UserResource(User::create($request->mappedAttributes()));
-         } catch (AuthorizationException $ex) {
-             return $this->error('You are not authorized to create that resource', 403);
-         }
+        return $this->error('You are not authorized to create that resource', 403);
     }
 
     /**
@@ -60,16 +56,14 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            // policy
-            $this->isAble('update', $user);
+            if ($this->isAble('update', $user)) {
+                $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-            $user->update($request->mappedAttributes());
-
-            return new UserResource($user);
+            return $this->error('You are not authorized to update that resource', 403);
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 403);
         }
     }
 
@@ -82,16 +76,14 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            // policy
-            $this->isAble('replace', $user);
+            if ($this->isAble('replace', $user)) {
+                $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
 
-            $user->update($request->mappedAttributes());
-
-            return new UserResource($user);
+            return $this->error('You are not authorized to update that resource', 403);
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to replace that resource', 403);
         }
     }
 
@@ -103,16 +95,14 @@ class UserController extends ApiController
         try {
             $user = User::findOrFail($user_id);
 
-            // policy
-            $this->isAble('delete', $user);
+            if ($this->isAble('delete', $user)) {
+                $user->delete();
+                return $this->ok('User successfully deleted');
+            }
 
-            $user->delete();
-
-            return $this->ok('User successfully deleted');
+            return $this->error('You are not authorized to delete that resource', 403);
         } catch (ModelNotFoundException $exception) {
             return $this->error('User cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to delete that resource', 403);
         }
     }
 }
